CopyComponent and CopyComponentIfExists as a variadic template functions
CopyComponent and CopyComponentIfExists as a variadic template functions. It doesn't need a lambda but I perfer that to some of the other methods.

I also attempted to use an empty struct to hold Groups of Component template arguments. You can pass those in as a function overload.

I'm not an expert they are kinds of things I've seen before.