Add Premake version auto lookup
Update the SetupPremake.py to download the latest available Premake
release by doing a GET request for release metadata in JSON format
and getting the first tag name. It strips out the leading v from the
semantic versioning. If the request fails it will default to the
5.0.0-beta1.

Credit for the GET request function goes to @nhathuy13598. I modified
their code by adding a comment and removing the classmethod directive,
as the GetLatestPremakeVersion function needed to be declared before
the class variables. I elected to define this function outside the
PremakeConfiguration class to keep the class variables at the top of
the file.

Closes TheCherno/Hazel#545